Add a basic, optional cache layer (#17)
Part of #5.

This change adds a basic caching layer which uses the Google Cloud Datastore to store cached content to serve. Summary rationale for choosing datastore over a memcached solution like redis is that this caching layer is unique in that it is trying to avoid expensive rendering costs instead of sub-ms response times. Datastore retrieval times are slower, but allow us to implement a perfect cache with large storage backing it.